Replace flat Node array with ndarray Array
This removes the previous, naieve array index juggling needed when
modifying a Schematic (merge, fill, insert_layer), simplifying the code
by a lot.  It's also much faster overall.

On top of that, it will now be easy to split a large Schematic into
partitions and save them separately, as Luanti eats a lot of memory (>
16GB) when loading huge Schematics (like 2048x256x2048)